requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
pillow>=9.0.0
python-dotenv>=0.19.0
pandas>=2.0.0
